Couple y-axis zoom of both sub plots (library / query) JCheckboxMenuItem
