+++ src/main/java/net/sf/mzmine/modules/visualization/spectra/simplespectra/spectraidentification/spectraldatabase/SpectraIdentificationResultsWindow.java
@@ -20,46 +20,24 @@ package net.sf.mzmine.modules.visualization.spectra.simplespectra.spectraidentif
 
 import java.awt.BorderLayout;
 import java.awt.Color;
-import java.awt.Component;
 import java.awt.Dimension;
 import java.awt.Font;
 import java.awt.GridLayout;
-import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.logging.Level;
+import java.util.Objects;
 import java.util.logging.Logger;
-import javax.swing.BorderFactory;
-import javax.swing.Box;
-import javax.swing.BoxLayout;
-import javax.swing.JComponent;
+import javax.swing.JCheckBoxMenuItem;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JMenuBar;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
-import javax.swing.ScrollPaneConstants;
 import javax.swing.SwingConstants;
 import javax.swing.SwingUtilities;
-import org.openscience.cdk.DefaultChemObjectBuilder;
-import org.openscience.cdk.exception.CDKException;
-import org.openscience.cdk.exception.InvalidSmilesException;
-import org.openscience.cdk.inchi.InChIGeneratorFactory;
-import org.openscience.cdk.inchi.InChIToStructure;
-import org.openscience.cdk.interfaces.IAtomContainer;
-import org.openscience.cdk.smiles.SmilesParser;
-import net.miginfocom.swing.MigLayout;
 import net.sf.mzmine.desktop.impl.WindowsMenu;
-import net.sf.mzmine.framework.CustomTextPane;
-import net.sf.mzmine.framework.ScrollablePanel;
-import net.sf.mzmine.modules.visualization.molstructure.Structure2DComponent;
-import net.sf.mzmine.modules.visualization.spectra.simplespectra.mirrorspectra.MirrorScanWindow;
-import net.sf.mzmine.util.ColorScaleUtil;
-import net.sf.mzmine.util.components.MultiLineLabel;
-import net.sf.mzmine.util.spectraldb.entry.DBEntryField;
-import net.sf.mzmine.util.spectraldb.entry.SpectralDBEntry;
 import net.sf.mzmine.util.spectraldb.entry.SpectralDBPeakIdentity;
 
 /**
@@ -69,26 +47,14 @@ import net.sf.mzmine.util.spectraldb.entry.SpectralDBPeakIdentity;
  */
 public class SpectraIdentificationResultsWindow extends JFrame {
   private final Logger logger = Logger.getLogger(this.getClass().getName());
-  private static final long serialVersionUID = 1L;
-  public static final int META_WIDTH = 500;
-  public static final int ENTRY_HEIGHT = 500;
 
-  // colors
-  public static final double MIN_COS_COLOR_VALUE = 0.5;
-  public static final double MAX_COS_COLOR_VALUE = 1.0;
-  // min color is a darker red
-  // max color is a darker green
-  public static final Color MAX_COS_COLOR = new Color(0x388E3C);
-  public static final Color MIN_COS_COLOR = new Color(0xE30B0B);
-
-  private JPanel pnGrid;
-  private Font titleFont = new Font("Dialog", Font.BOLD, 18);
-  private Font scoreFont = new Font("Dialog", Font.BOLD, 30);
   private Font headerFont = new Font("Dialog", Font.BOLD, 16);
-  private static final DecimalFormat COS_FORM = new DecimalFormat("0.000");
+  private JPanel pnGrid;
   private JScrollPane scrollPane;
   private List<SpectralDBPeakIdentity> totalMatches;
-  private Map<SpectralDBPeakIdentity, JPanel> matchPanels;
+  private Map<SpectralDBPeakIdentity, SpectralMatchPanel> matchPanels;
+  // couple y zoom (if one is changed - change the other in a mirror plot)
+  private boolean isCouplingZoomY;
 
   public SpectraIdentificationResultsWindow() {
     setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
@@ -98,7 +64,7 @@ public class SpectraIdentificationResultsWindow extends JFrame {
 
     pnGrid = new JPanel();
     // any number of rows
-    pnGrid.setLayout(new GridLayout(0, 1, 0, 25));
+    pnGrid.setLayout(new GridLayout(0, 1, 0, 0));
 
     pnGrid.setBackground(Color.WHITE);
     pnGrid.setAutoscrolls(false);
@@ -112,6 +78,10 @@ public class SpectraIdentificationResultsWindow extends JFrame {
     // Add the Windows menu
     JMenuBar menuBar = new JMenuBar();
     menuBar.add(new WindowsMenu());
+    JCheckBoxMenuItem cbCoupleZoomY = new JCheckBoxMenuItem("Couple y-zoom");
+    cbCoupleZoomY.setSelected(true);
+    cbCoupleZoomY.addItemListener(e -> setCoupleZoomY(cbCoupleZoomY.isSelected()));
+    menuBar.add(cbCoupleZoomY);
     setJMenuBar(menuBar);
 
     scrollPane = new JScrollPane(pnGrid);
@@ -122,6 +92,7 @@ public class SpectraIdentificationResultsWindow extends JFrame {
 
     totalMatches = new ArrayList<>();
     matchPanels = new HashMap<>();
+    setCoupleZoomY(true);
 
     setVisible(true);
     setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
@@ -129,184 +100,13 @@ public class SpectraIdentificationResultsWindow extends JFrame {
     repaint();
   }
 
-  /**
-   * Creates a new panel for the library match
-   * 
-   * @param scan
-   * @param hit
-   * @return
-   */
-  private JPanel createPanel(SpectralDBPeakIdentity hit) {
-    JPanel panel = new JPanel(new BorderLayout());
+  public void setCoupleZoomY(boolean selected) {
+    isCouplingZoomY = selected;
 
-    JPanel spectrumPanel = new JPanel(new BorderLayout());
-
-    // set meta data from identity
-    JPanel metaDataPanel = new JPanel();
-    metaDataPanel.setLayout(new BoxLayout(metaDataPanel, BoxLayout.Y_AXIS));
-
-    metaDataPanel.setBackground(Color.WHITE);
-
-    // add title
-    MigLayout l = new MigLayout("aligny center, wrap, insets 0 10 0 30", "[grow][]", "[grow]");
-    JPanel boxTitlePanel = new JPanel();
-    boxTitlePanel.setLayout(l);
-
-
-    double simScore = hit.getSimilarity().getScore();
-    Color gradientCol = ColorScaleUtil.getColor(MIN_COS_COLOR, MAX_COS_COLOR, MIN_COS_COLOR_VALUE,
-        MAX_COS_COLOR_VALUE, simScore);
-    boxTitlePanel.setBackground(gradientCol);
-    Box boxTitle = Box.createHorizontalBox();
-    // boxTitle.add(Box.createHorizontalGlue());
-
-    JPanel panelTitle = new JPanel(new GridLayout(1, 1));
-    panelTitle.setBackground(gradientCol);
-    String name = hit.getEntry().getField(DBEntryField.NAME).orElse("N/A").toString();
-    CustomTextPane title = new CustomTextPane(true);
-    title.setEditable(false);
-    title.setFont(headerFont);
-    title.setText(name);
-    title.setBackground(gradientCol);
-    title.setForeground(Color.WHITE);
-    panelTitle.add(title);
-
-    // score result
-    JPanel panelScore = new JPanel();
-    panelScore.setLayout(new BoxLayout(panelScore, BoxLayout.Y_AXIS));
-    JLabel score = new JLabel(COS_FORM.format(simScore));
-    score.setToolTipText("Cosine similarity of raw data scan (top, blue) and database scan: "
-        + COS_FORM.format(simScore));
-    score.setFont(scoreFont);
-    score.setForeground(Color.WHITE);
-    panelScore.setBackground(gradientCol);
-    panelScore.add(score);
-    boxTitlePanel.add(panelScore, "cell 1 0");
-    boxTitlePanel.add(panelTitle, "cell 0 0, growx, center");
-    boxTitle.add(boxTitlePanel);
-
-    // structure preview
-    IAtomContainer molecule;
-    JPanel preview2DPanel = new JPanel(new BorderLayout());
-    preview2DPanel.setPreferredSize(new Dimension(META_WIDTH, 150));
-    preview2DPanel.setMinimumSize(new Dimension(META_WIDTH, 150));
-    preview2DPanel.setMaximumSize(new Dimension(META_WIDTH, 150));
-    JComponent newComponent = null;
-
-    String inchiString = hit.getEntry().getField(DBEntryField.INCHI).orElse("N/A").toString();
-    String smilesString = hit.getEntry().getField(DBEntryField.SMILES).orElse("N/A").toString();
-
-    // check for INCHI
-    if (inchiString != "N/A") {
-      molecule = parseInChi(hit);
+    synchronized (matchPanels) {
+      matchPanels.values().stream().filter(Objects::nonNull)
+          .forEach(pn -> pn.setCoupleZoomY(selected));
     }
-    // check for smiles
-    else if (smilesString != "N/A") {
-      molecule = parseSmiles(hit);
-    } else
-      molecule = null;
-
-    // try to draw the component
-    if (molecule != null) {
-      try {
-        newComponent = new Structure2DComponent(molecule);
-      } catch (Exception e) {
-        String errorMessage = "Could not load 2D structure\n" + "Exception: ";
-        logger.log(Level.WARNING, errorMessage, e);
-        newComponent = new MultiLineLabel(errorMessage);
-      }
-      preview2DPanel.add(newComponent, BorderLayout.CENTER);
-      preview2DPanel.revalidate();
-
-      metaDataPanel.add(preview2DPanel);
-    }
-
-    // information on compound
-    JPanel panelCompounds =
-        extractMetaData("Compound information", hit.getEntry(), DBEntryField.COMPOUND_FIELDS);
-
-    // instrument info
-    JPanel panelInstrument =
-        extractMetaData("Instrument information", hit.getEntry(), DBEntryField.INSTRUMENT_FIELDS);
-
-    JPanel g1 = new JPanel(new GridLayout(1, 2, 4, 0));
-    g1.setBackground(Color.WHITE);
-    g1.add(panelCompounds);
-    g1.add(panelInstrument);
-    metaDataPanel.add(g1);
-
-    // database links
-    JPanel panelDB =
-        extractMetaData("Database links", hit.getEntry(), DBEntryField.DATABASE_FIELDS);
-
-    // // Other info
-    JPanel panelOther =
-        extractMetaData("Other information", hit.getEntry(), DBEntryField.OTHER_FIELDS);
-
-    JPanel g2 = new JPanel(new GridLayout(1, 2, 4, 0));
-    g2.setBackground(Color.WHITE);
-    g2.add(panelDB);
-    g2.add(panelOther);
-    metaDataPanel.add(g2);
-
-    // get mirror spectra window
-    MirrorScanWindow mirrorWindow = new MirrorScanWindow();
-    mirrorWindow.setScans(hit);
-
-    // fixed width panel
-    ScrollablePanel pn = new ScrollablePanel(new BorderLayout());
-    pn.setScrollableWidth(ScrollablePanel.ScrollableSizeHint.FIT);
-    pn.setScrollableHeight(ScrollablePanel.ScrollableSizeHint.STRETCH);
-    pn.add(metaDataPanel);
-
-
-    JScrollPane metaDataPanelScrollPane =
-        new JScrollPane(pn, ScrollPaneConstants.VERTICAL_SCROLLBAR_AS_NEEDED,
-            ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER);
-
-    spectrumPanel.add(mirrorWindow.getMirrorSpecrumPlot());
-
-    metaDataPanelScrollPane.setPreferredSize(new Dimension(META_WIDTH + 20, ENTRY_HEIGHT));
-    panel.setPreferredSize(new Dimension(0, ENTRY_HEIGHT));
-
-    panel.add(boxTitle, BorderLayout.NORTH);
-    panel.add(spectrumPanel, BorderLayout.CENTER);
-    panel.add(metaDataPanelScrollPane, BorderLayout.EAST);
-    panel.setBorder(BorderFactory.createLineBorder(Color.BLACK));
-
-    metaDataPanelScrollPane.revalidate();
-    pn.revalidate();
-    panel.revalidate();
-    panel.repaint();
-    return panel;
-  }
-
-  private JPanel extractMetaData(String title, SpectralDBEntry entry, DBEntryField[] other) {
-    JPanel panelOther = new JPanel();
-    panelOther.setLayout(new BoxLayout(panelOther, BoxLayout.Y_AXIS));
-    panelOther.setBackground(Color.WHITE);
-    panelOther.setAlignmentY(Component.TOP_ALIGNMENT);
-    panelOther.setAlignmentX(Component.TOP_ALIGNMENT);
-
-    for (DBEntryField db : other) {
-      Object o = entry.getField(db).orElse("N/A");
-      if (!o.equals("N/A")) {
-        CustomTextPane textPane = new CustomTextPane(true);
-        textPane.setText(db.toString() + ": " + o.toString());
-        panelOther.add(textPane);
-      }
-    }
-
-    JLabel otherInfo = new JLabel(title);
-    otherInfo.setFont(headerFont);
-    JPanel pn = new JPanel(new BorderLayout());
-    pn.setBackground(Color.WHITE);
-    pn.add(otherInfo, BorderLayout.NORTH);
-    pn.add(panelOther, BorderLayout.CENTER);
-    JPanel pn1 = new JPanel(new BorderLayout());
-    pn1.add(pn, BorderLayout.NORTH);
-    pn1.setBackground(Color.WHITE);
-    return pn1;
   }
 
   /**
@@ -319,7 +119,9 @@ public class SpectraIdentificationResultsWindow extends JFrame {
     if (!totalMatches.contains(match)) {
       // add
       totalMatches.add(match);
-      matchPanels.put(match, createPanel(match));
+      SpectralMatchPanel pn = new SpectralMatchPanel(match);
+      pn.setCoupleZoomY(isCouplingZoomY);
+      matchPanels.put(match, pn);
 
       // sort and show
       sortTotalMatches();
@@ -338,7 +140,9 @@ public class SpectraIdentificationResultsWindow extends JFrame {
       if (!totalMatches.contains(match)) {
         // add
         totalMatches.add(match);
-        matchPanels.put(match, createPanel(match));
+        SpectralMatchPanel pn = new SpectralMatchPanel(match);
+        pn.setCoupleZoomY(isCouplingZoomY);
+        matchPanels.put(match, pn);
       }
     }
     // sort and show
@@ -372,7 +176,7 @@ public class SpectraIdentificationResultsWindow extends JFrame {
   public void renewLayout() {
     SwingUtilities.invokeLater(() -> {
       // any number of rows
-      JPanel pnGrid = new JPanel(new GridLayout(0, 1, 0, 25));
+      JPanel pnGrid = new JPanel(new GridLayout(0, 1, 0, 5));
       pnGrid.setBackground(Color.WHITE);
       pnGrid.setAutoscrolls(false);
       // add all panel in order
@@ -392,43 +196,4 @@ public class SpectraIdentificationResultsWindow extends JFrame {
       this.pnGrid = pnGrid;
     });
   }
-
-  private IAtomContainer parseInChi(SpectralDBPeakIdentity hit) {
-    String inchiString = hit.getEntry().getField(DBEntryField.INCHI).orElse("N/A").toString();
-    InChIGeneratorFactory factory;
-    IAtomContainer molecule;
-    if (inchiString != "N/A") {
-      try {
-        factory = InChIGeneratorFactory.getInstance();
-        // Get InChIToStructure
-        InChIToStructure inchiToStructure =
-            factory.getInChIToStructure(inchiString, DefaultChemObjectBuilder.getInstance());
-        molecule = inchiToStructure.getAtomContainer();
-        return molecule;
-      } catch (CDKException e) {
-        String errorMessage = "Could not load 2D structure\n" + "Exception: ";
-        logger.log(Level.WARNING, errorMessage, e);
-        return null;
-      }
-    } else
-      return null;
-  }
-
-  private IAtomContainer parseSmiles(SpectralDBPeakIdentity hit) {
-    SmilesParser smilesParser = new SmilesParser(DefaultChemObjectBuilder.getInstance());
-    String smilesString = hit.getEntry().getField(DBEntryField.SMILES).orElse("N/A").toString();
-    IAtomContainer molecule;
-    if (smilesString != "N/A") {
-      try {
-        molecule = smilesParser.parseSmiles(smilesString);
-        return molecule;
-      } catch (InvalidSmilesException e1) {
-        String errorMessage = "Could not load 2D structure\n" + "Exception: ";
-        logger.log(Level.WARNING, errorMessage, e1);
-        return null;
-      }
-    } else
-      return null;
-  }
-
 }
